fix: cache topics per tenant
