+++ src/main/java/de/caritas/cob/agencyservice/api/service/TopicCachingService.java
@@ -1,0 +1,37 @@
+package de.caritas.cob.agencyservice.api.service;
+
+import de.caritas.cob.agencyservice.api.service.securityheader.SecurityHeaderSupplier;
+import de.caritas.cob.agencyservice.config.CacheManagerConfig;
+import de.caritas.cob.agencyservice.config.apiclient.TopicServiceApiControllerFactory;
+import de.caritas.cob.agencyservice.topicservice.generated.ApiClient;
+import de.caritas.cob.agencyservice.topicservice.generated.web.TopicControllerApi;
+import de.caritas.cob.agencyservice.topicservice.generated.web.model.TopicDTO;
+import java.util.List;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class TopicCachingService {
+
+  private final @NonNull TopicServiceApiControllerFactory topicServiceApiControllerFactory;
+  private final @NonNull SecurityHeaderSupplier securityHeaderSupplier;
+  private final @NonNull TenantHeaderSupplier tenantHeaderSupplier;
+
+  @Cacheable(cacheNames = CacheManagerConfig.TOPICS_CACHE)
+  public List<TopicDTO> getAllTopics(Long tenantId) {
+    // tenantId parameter is just used for caching per tenant, it is not required for the actual implementation, because it will be added by the tenantHeaderSupplier
+    TopicControllerApi controllerApi = topicServiceApiControllerFactory.createControllerApi();
+    addDefaultHeaders(controllerApi.getApiClient());
+    return controllerApi.getAllTopics();
+  }
+
+  private void addDefaultHeaders(ApiClient apiClient) {
+    var headers = this.securityHeaderSupplier.getKeycloakAndCsrfHttpHeaders();
+    tenantHeaderSupplier.addTenantHeader(headers);
+    headers.forEach((key, value) -> apiClient.addDefaultHeader(key, value.iterator().next()));
+  }
+
+}
