+++ src/main/java/de/caritas/cob/agencyservice/api/service/TopicService.java
@@ -1,35 +1,22 @@
 package de.caritas.cob.agencyservice.api.service;
 
-import de.caritas.cob.agencyservice.api.service.securityheader.SecurityHeaderSupplier;
-import de.caritas.cob.agencyservice.config.CacheManagerConfig;
-import de.caritas.cob.agencyservice.config.apiclient.TopicServiceApiControllerFactory;
-import de.caritas.cob.agencyservice.topicservice.generated.web.TopicControllerApi;
+import de.caritas.cob.agencyservice.api.tenant.TenantContext;
+
 import java.util.List;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import de.caritas.cob.agencyservice.topicservice.generated.web.model.TopicDTO;
-import de.caritas.cob.agencyservice.topicservice.generated.ApiClient;
 
 @Service
 @RequiredArgsConstructor
 public class TopicService {
 
-  private final @NonNull TopicServiceApiControllerFactory topicServiceApiControllerFactory;
-  private final @NonNull SecurityHeaderSupplier securityHeaderSupplier;
-  private final @NonNull TenantHeaderSupplier tenantHeaderSupplier;
+  private final @NonNull TopicCachingService topicCachingService;
+
 
   public List<TopicDTO> getAllTopics() {
-    TopicControllerApi controllerApi = topicServiceApiControllerFactory.createControllerApi();
-    addDefaultHeaders(controllerApi.getApiClient());
-    return controllerApi.getAllTopics();
-  }
-
-  private void addDefaultHeaders(ApiClient apiClient) {
-    var headers = this.securityHeaderSupplier.getKeycloakAndCsrfHttpHeaders();
-    tenantHeaderSupplier.addTenantHeader(headers);
-    headers.forEach((key, value) -> apiClient.addDefaultHeader(key, value.iterator().next()));
+    return topicCachingService.getAllTopics(TenantContext.getCurrentTenant());
   }
 
 }
